pyproject.toml
[tool.poetry]
name = "test-ORBIDI"
version = "0.1.0"
description = "Descripción del proyecto"
authors = ["Sebastian Noreña Marquez <sebastian.norena.marquez@gmail.com>"]

[tool.poetry.group.dev.dependencies]
pytest-mock = "^3.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.68.0"
pydantic = "^1.8.2"
sqlalchemy = "^1.4.22"
python-dotenv = "^0.19.0"
requests = "^2.26.0"
psycopg2-binary = "^2.9.1"
uvicorn = "^0.22.0"
email-validator = "*"
requests-mock = "^1.11.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"

[tool.poetry.scripts]
lint = "flake8 --exclude=venv,.venv,.git,__pycache__,dist,build,tests"